# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API маркетплейса Клевер (микросервисная архитектура).\n- Auth Service       (HTTP :8001, gRPC :9001) — авторизация, профиль, пользователи\n- Support Service    (HTTP :8002)             — техподдержка\n- Catalog Service    (HTTP :8004, gRPC :9004) — объявления, категории, избранное, просмотры\n- Commerce Service   (HTTP :8006)             — корзина, чаты, заказы\nВсе запросы идут через API Gateway (nginx :8080).",
        "title": "Clover API",
        "contact": {},
        "version": "1.0"
    },
    "host": "clover-go.ru",
    "basePath": "/api/v1",
    "paths": {
...
        "/wallet": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Получить баланс кошелька",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WalletResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallet/payments/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Получить статус платежа",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID платежа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaymentStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallet/topup": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Пополнить кошелёк",
                "parameters": [
                    {
                        "description": "Сумма в рублях и idempotency_key",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TopupWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TopupWalletResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_AMOUNT / INVALID_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallet/transactions": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "История операций по кошельку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "лимит (по умолчанию 20, максимум 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID последнего элемента предыдущей страницы",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WalletTransactionListResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallet/yookassa/webhook": {
            "post": {
                "description": "Принимает уведомления от ЮКассы. Доступ ограничен по IP. Всегда отвечает 200, чтобы провайдер не ретраил — расхождения добираются reconciler'ом.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Webhook ЮКассы об изменении статуса платежа",
                "parameters": [
                    {
                        "description": "Notification payload ЮКассы",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaymentStatusResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "confirmation_url": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.TopupWalletRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "string"
                }
            }
        },
        "dto.TopupWalletResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "confirmation_url": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.WalletResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
        "dto.WalletTransactionListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.WalletTransactionResponse"
                    }
                },
                "next_cursor": {
                    "type": "integer"
                }
            }
        },
        "dto.WalletTransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reference_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT session token",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        },
        "CsrfCookieAuth": {
            "description": "CSRF token stored in cookies",
            "type": "apiKey",
            "name": "csrf_token",
            "in": "cookie"
        },
        "CsrfHeaderAuth": {
            "description": "CSRF token required in header",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}